Animation leakage issue fixed
Issue was due to fragments. Because of how fragment works ideologically, using it means that there is a miscommunication between native view manager and then the vdom(engine). The reason is that Fragment is, i made to represent something not meant to exist on the native layer, this is very useful for lower-level apis that don't want to interact much with native components but just deceive the vdom to represent nodes in a particular pattern. Using this on very high level can cause data structure issues at the native side